Restart with trial of new structure
