+++ HeadUnit/GearSelector/.qmlls.ini
@@ -1,0 +1,5 @@
+[General]
+buildDir="C:/Users/aakas/Documents/GitHub/DES_Head-Unit/HeadUnit/GearSelector/build/Debug"
+no-cmake-calls=false
+docDir=C:/Qt/Docs/Qt-6.9.2
+importPaths="C:/Qt/6.9.2/mingw_64/qml"
