+++ HeadUnit/GearSelector/gs_handler.cpp
@@ -1,0 +1,5 @@
+#include "gs_handler.h"
+
+GS_Handler::GS_Handler(QObject *parent)
+    : QObject{parent}
+{}
