--- HeadUnit/services/media-service/main.cpp
@@ -1,5 +1,0 @@
-#include <QCoreApplication>
-int main(int argc, char *argv[]) {
-    QCoreApplication app(argc, argv);
-    return app.exec();
-}
